requirements.txt
llama-index
chromadb
streamlit

openai

pypdf2
python-docx
python-pptx

python-dotenv
pandas
numpy

pytest
pytest-asyncio

requests
pathlib2